requirements.txt
selenium==4.27.0
python-dotenv==1.0.1
pyperclip==1.8.2